organize hubs endpoints under /hubs
