+++ backend/src/Designer/Hubs/HubsEndpointExtensions.cs
@@ -1,4 +1,5 @@
 using Altinn.Studio.Designer.Hubs.EntityUpdate;
+using Altinn.Studio.Designer.Hubs.Preview;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
 
@@ -9,8 +10,8 @@ public static class HubsEndpointExtensions
     public static void MapHubs(this IEndpointRouteBuilder endpoints)
     {
         // all hubs should be changed to have endpoint /hubs/hub-name
-        endpoints.MapHub<PreviewHub>("/previewHub");
-        endpoints.MapHub<SyncHub.SyncHub>("/sync-hub");
+        endpoints.MapHub<PreviewHub>("/hubs/preview");
+        endpoints.MapHub<SyncHub.SyncHub>("/hubs/sync");
         endpoints.MapHub<EntityUpdatedHub>("/hubs/entity-updated");
     }
 }
