+++ backend/src/Designer/Hubs/Preview/IPreviewClient.cs
@@ -1,0 +1,8 @@
+using System.Threading.Tasks;
+
+namespace Altinn.Studio.Designer.Hubs.Preview;
+
+public interface IPreviewClient
+{
+    Task ReceiveMessage(string message);
+}
