+++ backend/src/Designer/Hubs/Preview/PreviewHub.cs
@@ -5,10 +5,10 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.Logging;
 
-namespace Altinn.Studio.Designer.Hubs;
+namespace Altinn.Studio.Designer.Hubs.Preview;
 
 [Authorize]
-public class PreviewHub : Hub
+public class PreviewHub : Hub<IPreviewClient>
 {
     private readonly ILogger _logger;
     private readonly IHttpContextAccessor _httpContextAccessor;
@@ -19,18 +19,18 @@ public class PreviewHub : Hub
         _httpContextAccessor = httpContextAccessor;
     }
 
-    public override Task OnConnectedAsync()
+    public override async Task OnConnectedAsync()
     {
         string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
         string connectionId = Context.ConnectionId;
-        Groups.AddToGroupAsync(connectionId, developer);
-        return base.OnConnectedAsync();
+        await Groups.AddToGroupAsync(connectionId, developer);
+        await base.OnConnectedAsync();
     }
 
     public async Task SendMessage(string message)
     {
         string developer = AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext);
         _logger.LogInformation("Message received from client: {MessageFromClient}", message);
-        await Clients.Group(developer).SendAsync("ReceiveMessage", message);
+        await Clients.Group(developer).ReceiveMessage(message);
     }
 }
