+++ charts/altinn-loadbalancer/templates/configmap.yaml
@@ -95,26 +95,6 @@ data:
           proxy_set_header    X-Forwarded-Proto $scheme;
         }
 
-        location /previewHub {
-          proxy_pass            http://designer/previewHub;
-          proxy_http_version  1.1;
-          proxy_set_header    Upgrade $http_upgrade;
-          proxy_set_header    Connection "Upgrade";
-          proxy_set_header    Host {{ $host }};
-          proxy_set_header 	  X-Forwarded-For $remote_addr;
-          proxy_set_header    X-Forwarded-Proto $scheme;
-        }
-
-        location /sync-hub {
-          proxy_pass          http://designer/sync-hub;
-          proxy_http_version  1.1;
-          proxy_set_header    Upgrade $http_upgrade;
-          proxy_set_header    Connection "Upgrade";
-          proxy_set_header    Host {{ $host }};
-          proxy_set_header 	  X-Forwarded-For $remote_addr;
-          proxy_set_header    X-Forwarded-Proto $scheme;
-        }
-
         location /hubs/ {
           proxy_pass            http://designer/hubs/;
           proxy_http_version    1.1;
