+++ development/load-balancer/templates/nginx.conf.conf
@@ -99,20 +99,6 @@ http {
 			}
 		}
 
-    location ~ ^/(previewHub|sync-hub) {
-      proxy_http_version 1.1;
-      proxy_set_header Upgrade $http_upgrade;
-      proxy_set_header Connection "Upgrade";
-      if ($dev_backend) {
-        add_header X-Dashboard-Source dotnetPaths;
-        proxy_pass http://host.docker.internal:5000;
-      }
-      if ($dev_backend != 1) {
-        add_header X-Dashboard-Source dockerPaths;
-        proxy_pass http://designer;
-      }
-    }
-
     location ~ ^/hubs/ {
       proxy_http_version  1.1;
       proxy_set_header    Upgrade $http_upgrade;
